Workaround for BSD sed -i issue
BSD sed’s -i option is considered harmful. It leaves dotfiles behind and
does not clean them up. For BSD sed I manually output to a temp file
then overwrite the original.

For non BSD sed it uses the -i (--in-place) option and runs the command
in the background in a subshell to (hopefully) speed things up.