Ensure exit status is captured correctly
For some reason, in Ubuntu 16.04 (but not 18.04), I have to capture the
value of $? (the last command’s exit status) at the top of
prompt_command, even before any comments, or $? will be set to a
different value. Apparently the comment lines are giving off a return
value of 0.

The odd part is that I am sure I had comments above that variable
assignment before without this problems in Ubuntu 16.04.