Revert previous commit
Disregard the previous commit comment. In Ubuntu 16.04 the exit status
is being set back to 0 by an extra call to prompt_handle_debug between
when the command is entered and when prompt_command is called. I still
do not know why this is happening in Ubuntu 16.04 and not 18.04, nor do
I know why it has only started happening recently.

I could capture $? at the top of prompt_handle_debug, then set a global
variable from it when prompt_debug_marks is at the correct value, but it
would still be different between Ubuntu 16.04 and 18.04. It’s currently
fine in 18.04, so I will leave it as is for now. If an extra
prompt_handle_debug call shows up in 18.04 I will have to implement
that.