Properly fix exit status display and add warning to README
The problem I was seeing with the exit status display was from me adding
something to my PROMPT_COMMAND environment variable that was not called
inside a subshell like z.

I cannot make jm-shell entirely invulnerable to this, but I changed it
so that you can add other command or function calls to your
PROMPT_COMMAND so long as they are either in a subshell, or they do not
come after the call to jm-shell’s `prompt_command` function.

I added a troubleshooting section to the README that covers this
starting with the problems it can cause.